Refactor into individual Cmd objects
