+++ src/lib.rs
@@ -4,13 +4,7 @@ extern crate prettytable;
 #[macro_use]
 extern crate lazy_static;
 
-pub mod cmd_balance;
-pub mod cmd_create;
-pub mod cmd_hotspots;
-pub mod cmd_htlc;
-pub mod cmd_info;
-pub mod cmd_pay;
-pub mod cmd_verify;
+pub mod cmd;
 pub mod keypair;
 pub mod mnemonic;
 pub mod result;
